Move Copasi Optimization Solver to Client
Instead of the copasi optimization solver being within core, its now within client.
